// frontend/src/clients/analytics/types.gen.ts
export type ChartData = {
    /**
     * Text1
     */
    text1: string;
    /**
     * Value1
     */
    value1: string | number | number;
    /**
     * Documentcount
     */
    documentCount: number;
    /**
     * Sentencecount
     */
    sentenceCount: number;
    /**
     * Mentioncount
     */
    mentionCount: number;
    /**
     * Text2
     */
    text2?: string;
    /**
     * Value2
     */
    value2?: string | number | number;
};
// ...
export type CountType = 'document' | 'sentence' | 'mention';

/**
 * DataCategory
 */
export type DataCategory = 'string' | 'number' | 'boolean' | 'date' | 'none';